Unify the implementation under FuzzyAction and create tests
