+++ src/test/kotlin/com/mituuz/fuzzier/FuzzyMoverTest.kt
@@ -19,12 +19,10 @@ class FuzzyMoverTest {
     private var fuzzyMover: FuzzyMover
     private var testApplicationManager: TestApplicationManager
     private val testUtil: TestUtil = TestUtil()
-    private var settings: FuzzierSettingsService.State
 
     init {
         testApplicationManager = TestApplicationManager.getInstance()
         fuzzyMover = FuzzyMover()
-        settings  = service<FuzzierSettingsService>().state
     }
 
     @Test
@@ -71,62 +69,5 @@ class FuzzyMoverTest {
                 assertNull(targetFile)
             }.join()
         }
-    }
-
-    @Test
-    fun `Check renderer filename only`() {
-        settings.filenameType = StringEvaluator.FilenameType.FILENAME_ONLY
-        val myFixture: CodeInsightTestFixture = testUtil.setUpProject(emptyList())
-        val project = myFixture.project
-        fuzzyMover.component = SimpleFinderComponent(project)
-        val renderer = fuzzyMover.getCellRenderer()
-        val container = FuzzyMatchContainer(1, "/src/asd", "asd")
-        val dummyList = JList<FuzzyMatchContainer>()
-        val component = renderer.getListCellRendererComponent(dummyList, container, 0, false, false) as JLabel
-        assertNotNull(component)
-        assertEquals("asd", component.text)
-    }
-
-    @Test
-    fun `Check renderer with path`() {
-        settings.filenameType = StringEvaluator.FilenameType.FILENAME_WITH_PATH
-        val myFixture: CodeInsightTestFixture = testUtil.setUpProject(emptyList())
-        val project = myFixture.project
-        fuzzyMover.component = SimpleFinderComponent(project)
-        val renderer = fuzzyMover.getCellRenderer()
-        val container = FuzzyMatchContainer(1, "/src/asd", "asd")
-        val dummyList = JList<FuzzyMatchContainer>()
-        val component = renderer.getListCellRendererComponent(dummyList, container, 0, false, false) as JLabel
-        assertNotNull(component)
-        assertEquals("asd (/src/asd)", component.text)
-    }
-
-    @Test
-    fun `Check renderer full path`() {
-        settings.filenameType = StringEvaluator.FilenameType.FILEPATH_ONLY
-        val myFixture: CodeInsightTestFixture = testUtil.setUpProject(emptyList())
-        val project = myFixture.project
-        fuzzyMover.component = SimpleFinderComponent(project)
-        val renderer = fuzzyMover.getCellRenderer()
-        val container = FuzzyMatchContainer(1, "/src/asd", "asd")
-        val dummyList = JList<FuzzyMatchContainer>()
-        val component = renderer.getListCellRendererComponent(dummyList, container, 0, false, false) as JLabel
-        assertNotNull(component)
-        assertEquals("/src/asd", component.text)
-    }
-
-    @Test
-    fun `Check renderer dir selector`() {
-        settings.filenameType = StringEvaluator.FilenameType.FILENAME_ONLY
-        val myFixture: CodeInsightTestFixture = testUtil.setUpProject(emptyList())
-        val project = myFixture.project
-        fuzzyMover.component = SimpleFinderComponent(project)
-        fuzzyMover.component.isDirSelector = true
-        val renderer = fuzzyMover.getCellRenderer()
-        val container = FuzzyMatchContainer(1, "/src/asd", "asd")
-        val dummyList = JList<FuzzyMatchContainer>()
-        val component = renderer.getListCellRendererComponent(dummyList, container, 0, false, false) as JLabel
-        assertNotNull(component)
-        assertEquals("/src/asd", component.text)
     }
 }
